+++ pom.xml
@@ -44,7 +44,6 @@
         <!-- JARs -->
         <module>tech1-framework-foundation</module>
         <module>tech1-framework-emails</module>
-        <module>tech1-framework-hardware</module>
         <module>tech1-framework-b2b-base-security-jwt</module>
         <module>tech1-framework-b2b-base-security-jwt-websockets</module>
         <module>tech1-framework-b2b-mongodb-security-jwt</module>
