+++ tech1-framework-b2b-base-security-jwt-websockets/src/main/java/io/tech1/framework/b2b/base/security/jwt/websockets/events/subscribers/HardwareMonitoringSubscriberWebsockets.java
@@ -2,8 +2,8 @@ package io.tech1.framework.b2b.base.security.jwt.websockets.events.subscribers;
 
 import io.tech1.framework.b2b.base.security.jwt.websockets.tasks.HardwareBackPressureTimerTask;
 import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
-import io.tech1.framework.hardware.monitoring.store.HardwareMonitoringStore;
-import io.tech1.framework.hardware.monitoring.subscribers.impl.BaseHardwareMonitoringSubscriber;
+import io.tech1.framework.foundation.services.hardware.store.HardwareMonitoringStore;
+import io.tech1.framework.foundation.services.hardware.subscribers.impl.BaseHardwareMonitoringSubscriber;
 import io.tech1.framework.foundation.incidents.events.publishers.IncidentPublisher;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
