+++ tech1-framework-b2b-base-security-jwt-websockets/src/main/java/io/tech1/framework/b2b/base/security/jwt/websockets/tasks/HardwareBackPressureTimerTask.java
@@ -4,7 +4,7 @@ import io.tech1.framework.b2b.base.security.jwt.sessions.SessionRegistry;
 import io.tech1.framework.b2b.base.security.jwt.websockets.tempate.WssMessagingTemplate;
 import io.tech1.framework.foundation.domain.concurrent.AbstractInfiniteTimerTask;
 import io.tech1.framework.foundation.domain.time.SchedulerConfiguration;
-import io.tech1.framework.hardware.monitoring.store.HardwareMonitoringStore;
+import io.tech1.framework.foundation.services.hardware.store.HardwareMonitoringStore;
 import io.tech1.framework.foundation.incidents.events.publishers.IncidentPublisher;
 import io.tech1.framework.foundation.domain.properties.ApplicationFrameworkProperties;
 import lombok.extern.slf4j.Slf4j;
