+++ tech1-framework-b2b-base-security-jwt-websockets/src/test/java/io/tech1/framework/b2b/base/security/jwt/websockets/events/subscribers/HardwareMonitoringSubscriberWebsocketsTest.java
@@ -2,7 +2,7 @@ package io.tech1.framework.b2b.base.security.jwt.websockets.events.subscribers;
 
 import io.tech1.framework.b2b.base.security.jwt.websockets.tasks.HardwareBackPressureTimerTask;
 import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
-import io.tech1.framework.hardware.monitoring.store.HardwareMonitoringStore;
+import io.tech1.framework.foundation.services.hardware.store.HardwareMonitoringStore;
 import io.tech1.framework.foundation.incidents.events.publishers.IncidentPublisher;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.AfterEach;
