+++ tech1-framework-b2b-mongodb-server/src/main/java/io/tech1/framework/b2b/mongodb/server/configurations/ApplicationTech1.java
@@ -5,10 +5,10 @@ import io.tech1.framework.b2b.base.security.jwt.configurations.ApplicationBaseSe
 import io.tech1.framework.b2b.mongodb.security.jwt.configurations.ApplicationMongo;
 import io.tech1.framework.b2b.mongodb.server.properties.ApplicationProperties;
 import io.tech1.framework.foundation.domain.base.PropertyId;
-import io.tech1.framework.hardware.configurations.ApplicationHardwareMonitoring;
-import io.tech1.framework.hardware.monitoring.store.HardwareMonitoringStore;
-import io.tech1.framework.hardware.monitoring.subscribers.HardwareMonitoringSubscriber;
-import io.tech1.framework.hardware.monitoring.subscribers.impl.BaseHardwareMonitoringSubscriber;
+import io.tech1.framework.foundation.configurations.ApplicationHardwareMonitoring;
+import io.tech1.framework.foundation.services.hardware.store.HardwareMonitoringStore;
+import io.tech1.framework.foundation.services.hardware.subscribers.HardwareMonitoringSubscriber;
+import io.tech1.framework.foundation.services.hardware.subscribers.impl.BaseHardwareMonitoringSubscriber;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
