+++ tech1-framework-foundation/src/main/java/io/tech1/framework/foundation/services/hardware/publishers/HardwareMonitoringPublisher.java
@@ -1,0 +1,7 @@
+package io.tech1.framework.foundation.services.hardware.publishers;
+
+import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
+
+public interface HardwareMonitoringPublisher {
+    void publishLastHardwareMonitoringDatapoint(EventLastHardwareMonitoringDatapoint event);
+}
