+++ tech1-framework-foundation/src/main/java/io/tech1/framework/foundation/services/hardware/store/HardwareMonitoringStore.java
@@ -1,0 +1,10 @@
+package io.tech1.framework.foundation.services.hardware.store;
+
+import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
+import io.tech1.framework.foundation.domain.hardware.monitoring.HardwareMonitoringWidget;
+
+public interface HardwareMonitoringStore {
+    HardwareMonitoringWidget getHardwareMonitoringWidget();
+    boolean containsOneElement();
+    void storeEvent(EventLastHardwareMonitoringDatapoint event);
+}
