+++ tech1-framework-foundation/src/main/java/io/tech1/framework/foundation/services/hardware/subscribers/impl/BaseHardwareMonitoringSubscriber.java
@@ -1,0 +1,20 @@
+package io.tech1.framework.foundation.services.hardware.subscribers.impl;
+
+import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
+import io.tech1.framework.foundation.domain.pubsub.AbstractEventSubscriber;
+import io.tech1.framework.foundation.services.hardware.store.HardwareMonitoringStore;
+import io.tech1.framework.foundation.services.hardware.subscribers.HardwareMonitoringSubscriber;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@AllArgsConstructor
+public class BaseHardwareMonitoringSubscriber extends AbstractEventSubscriber implements HardwareMonitoringSubscriber {
+
+    private final HardwareMonitoringStore hardwareMonitoringStore;
+
+    @Override
+    public void onLastHardwareMonitoringDatapoint(EventLastHardwareMonitoringDatapoint event) {
+        this.hardwareMonitoringStore.storeEvent(event);
+    }
+}
