--- tech1-framework-hardware/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>io.tech1.framework</groupId>
-        <artifactId>tech1-framework</artifactId>
-        <version>3.0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>tech1-framework-hardware</artifactId>
-
-    <dependencies>
-        <!-- Tech1: Framework -->
-        <dependency>
-            <groupId>io.tech1.framework</groupId>
-            <artifactId>tech1-framework-foundation</artifactId>
-            <version>${project.version}</version>
-            <scope>compile</scope>
-        </dependency>
-    </dependencies>
-</project>
