--- tech1-framework-hardware/src/main/java/io/tech1/framework/hardware/monitoring/publishers/HardwareMonitoringPublisher.java
@@ -1,7 +1,0 @@
-package io.tech1.framework.hardware.monitoring.publishers;
-
-import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
-
-public interface HardwareMonitoringPublisher {
-    void publishLastHardwareMonitoringDatapoint(EventLastHardwareMonitoringDatapoint event);
-}
