--- tech1-framework-hardware/src/main/java/io/tech1/framework/hardware/monitoring/store/HardwareMonitoringStore.java
@@ -1,10 +1,0 @@
-package io.tech1.framework.hardware.monitoring.store;
-
-import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
-import io.tech1.framework.foundation.domain.hardware.monitoring.HardwareMonitoringWidget;
-
-public interface HardwareMonitoringStore {
-    HardwareMonitoringWidget getHardwareMonitoringWidget();
-    boolean containsOneElement();
-    void storeEvent(EventLastHardwareMonitoringDatapoint event);
-}
