--- tech1-framework-hardware/src/main/java/io/tech1/framework/hardware/monitoring/subscribers/HardwareMonitoringSubscriber.java
@@ -1,9 +1,0 @@
-package io.tech1.framework.hardware.monitoring.subscribers;
-
-import io.tech1.framework.foundation.domain.events.hardware.EventLastHardwareMonitoringDatapoint;
-import org.springframework.context.event.EventListener;
-
-public interface HardwareMonitoringSubscriber {
-    @EventListener
-    void onLastHardwareMonitoringDatapoint(EventLastHardwareMonitoringDatapoint event);
-}
