--- tech1-framework-hardware/src/test/resources/logback-test.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration scan="true">
-    <statusListener class="ch.qos.logback.core.status.NopStatusListener" />
-
-    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-        <layout class="ch.qos.logback.classic.PatternLayout">
-            <Pattern>
-                %d{ISO8601} %5p --- [%15.15t] %-40.40logger{39} : %m%n
-            </Pattern>
-        </layout>
-    </appender>
-
-    <logger name="ch.qos.logback" level="warn" />
-    <logger name="com.ulisesbocchio" level="warn" />
-    <logger name="org.apache" level="error" />
-    <logger name="org.mongodb" level="warn" />
-    <logger name="org.springframework" level="warn" />
-
-    <root level="info">
-        <appender-ref ref="CONSOLE" />
-    </root>
-
-</configuration>
